requirements.txt
discord.py
discord.py[voice]
requests
youtube_dl
pyyaml